+++ app/models/journal.rb
@@ -6,11 +6,6 @@ class Journal < ActiveRecord::Base
   has_many :discussion_topics, through: :papers, inverse_of: :journal
   has_many :letter_templates
 
-  # Old Roles and Permissions
-  has_many :old_roles, inverse_of: :journal
-  has_many :user_roles, through: :old_roles
-  has_many :users, through: :user_roles
-
   has_many :manuscript_manager_templates, dependent: :destroy
   has_many :journal_task_types, inverse_of: :journal, dependent: :destroy
 
@@ -25,7 +20,6 @@ class Journal < ActiveRecord::Base
 
   after_create :setup_defaults
   before_destroy :confirm_no_papers, prepend: true
-  before_destroy :destroy_roles
 
   mount_uploader :logo, LogoUploader
 
@@ -74,16 +68,8 @@ class Journal < ActiveRecord::Base
     all.flat_map(&:staff_admins)
   end
 
-  def admins
-    users.merge(OldRole.admins)
-  end
-
   def staff_admins
     User.with_role(staff_admin_role, assigned_to: self)
-  end
-
-  def reviewers
-    users.merge(OldRole.reviewers)
   end
 
   def logo_url
@@ -98,10 +84,6 @@ class Journal < ActiveRecord::Base
     # We ordering by the oldest articles first to have 'Research Article'
     # to float to the top of the article drop down list
     manuscript_manager_templates.order('id asc').pluck(:paper_type)
-  end
-
-  def valid_old_roles
-    PaperRole::ALL_ROLES | old_roles.map(&:name)
   end
 
   # Try to block other services from directly updating last_doi_issued to avoid
@@ -131,16 +113,8 @@ class Journal < ActiveRecord::Base
 
   def setup_defaults
     # TODO: remove these from being a callback (when we aren't using rails_admin)
-    JournalServices::CreateDefaultRoles.call(self)
     JournalServices::CreateDefaultTaskTypes.call(self)
     JournalServices::CreateDefaultManuscriptManagerTemplates.call(self)
-  end
-
-  def destroy_roles
-    # old_roles that are marked as 'required' are prevented from being destroyed, so you cannot use
-    # a dependent_destroy on the AR relationship.
-    mark_for_destruction
-    old_roles.destroy_all
   end
 
   def confirm_no_papers