+++ client/tests/integration/paper-index-test.js
@@ -32,8 +32,7 @@ module('Integration: PaperIndex', {
     TestHelper.mockFindAll('discussion-topic', 1);
     figureTaskId = 94139;
     records = paperWithTask('FigureTask', {
-      id: figureTaskId,
-      oldRole: "author"
+      id: figureTaskId
     });
     currentPaper = records[0], figureTask = records[1], journal = records[2], phase = records[3];
     nestedQuestion = Factory.createRecord('NestedQuestion', {
@@ -111,8 +110,7 @@ test('on paper.index as a participant on a task but not author of paper', functi
   records = paperWithTask('Task', {
     id: 1,
     type: 'AdHocTask',
-    title: 'ReviewMe',
-    oldRole: 'reviewer'
+    title: 'ReviewMe'
   });
   currentPaper = records[0], task = records[1], journal = records[2], litePaper = records[3], phase = records[4];
   paperPayload = Factory.createPayload('paper');
@@ -134,8 +132,7 @@ test('on paper.index as a participant on a task and author of paper', function(a
   assert.expect(1);
   records = paperWithTask('ReviseTask', {
     id: 1,
-    qualifiedType: "TahiStandardTasks::ReviseTask",
-    oldRole: 'author'
+    qualifiedType: 'TahiStandardTasks::ReviseTask'
   });
   currentPaper = records[0], task = records[1], journal = records[2], litePaper = records[3], phase = records[4];
   paperPayload = Factory.createPayload('paper');