+++ db/migrate/20161103203118_remove_paper_admin_task.rb
@@ -1,0 +1,3 @@
+class RemovePaperAdminTask < DataMigration
+  RAKE_TASK_UP = 'data:migrate:remove_paper_admin_task_journal_task_types'.freeze
+end