+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/competing_interests_task.rb
@@ -1,7 +1,7 @@
 module TahiStandardTasks
   class CompetingInterestsTask < ::Task
     include MetadataTask
-    DEFAULT_TITLE = 'Competing Interests'
-    DEFAULT_ROLE = 'author'
+    DEFAULT_TITLE = 'Competing Interests'.freeze
+    DEFAULT_ROLE_HINT = 'author'.freeze
   end
 end