+++ engines/tahi_standard_tasks/spec/models/revise_task_spec.rb
@@ -27,7 +27,7 @@ describe TahiStandardTasks::ReviseTask do
         expect(task.reload.completed).to be(false)
       end
 
-      it "updates the task's phase"do
+      it "updates the task's phase" do
         subject.setup_new_revision paper, phase
         expect(task.reload.phase_id).to be(phase.id)
       end
@@ -49,6 +49,5 @@ describe TahiStandardTasks::ReviseTask do
 
   describe '.restore_defaults' do
     it_behaves_like '<Task class>.restore_defaults update title to the default'
-    it_behaves_like '<Task class>.restore_defaults update old_role to the default'
   end
 end