+++ engines/tahi_standard_tasks/spec/models/upload_manuscript_task_spec.rb
@@ -3,6 +3,5 @@ require 'rails_helper'
 describe TahiStandardTasks::UploadManuscriptTask do
   describe '.restore_defaults' do
     it_behaves_like '<Task class>.restore_defaults update title to the default'
-    it_behaves_like '<Task class>.restore_defaults update old_role to the default'
   end
 end